WeAnimal-DogClassifier/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
Werkzeug==0.14.1
Keras==2.2.4
tensorflow==1.13.1
numpy==1.16.3
scikit-image==0.15.0
Pillow==6.0.0
gunicorn==19.9.0
requests==2.22.0
